Split tag.hpp into three files
