+++ projectfiles/VC14/wesnoth.vcxproj.filters
@@ -1010,7 +1010,13 @@
     <ClCompile Include="..\..\src\utils\name_generator_factory.cpp">
       <Filter>utils</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\src\serialization\tag.cpp">
+    <ClCompile Include="..\..\src\serialization\schema\key.cpp">
+      <Filter>serialization</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\serialization\schema\tag.cpp">
+      <Filter>serialization</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\serialization\schema\type.cpp">
       <Filter>serialization</Filter>
     </ClCompile>
     <ClCompile Include="..\..\src\tests\main.cpp">
@@ -2513,7 +2519,13 @@
     <ClInclude Include="..\..\src\utils\scope_exit.hpp">
       <Filter>utils</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\src\serialization\tag.hpp">
+    <ClInclude Include="..\..\src\serialization\schema\key.hpp">
+      <Filter>serialization</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\serialization\schema\tag.hpp">
+      <Filter>serialization</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\serialization\schema\type.hpp">
       <Filter>serialization</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\actions\advancement.hpp">
