+++ source_lists/libwesnoth_core
@@ -21,7 +21,9 @@ serialization/parser.cpp
 serialization/preprocessor.cpp
 serialization/schema_validator.cpp
 serialization/string_utils.cpp
-serialization/tag.cpp
+serialization/schema/key.cpp
+serialization/schema/tag.cpp
+serialization/schema/type.cpp
 serialization/tokenizer.cpp
 serialization/unicode.cpp
 serialization/validator.cpp
