+++ src/serialization/schema/tag.hpp
@@ -14,231 +14,24 @@
 
 /**
  * @file
- * This file contains objects "tag" and "key", which are used to store
- * information about tags and keys while annotation parsing.
+ * This file contains object "tag", which is used to store
+ * information about tags while annotation parsing.
  */
 
 #pragma once
 
-#include <algorithm>
-#include <iostream>
 #include <map>
-#include <sstream>
 #include <string>
 #include <vector>
 #include <queue>
 
-#include <boost/regex.hpp>
 #include <boost/iterator/iterator_facade.hpp>
 #include <boost/range/iterator.hpp>
 #include "config.hpp"
+#include "serialization/schema/key.hpp"
 
 namespace schema_validation
 {
-
-/**
- * Stores information about a schema type.
- * This is an abstract base class for several variants of schema type.
- */
-class class_type {
-protected:
-	std::string name_;
-public:
-	class_type() = delete;
-	explicit class_type(const std::string& name) : name_(name) {}
-	using ptr = std::shared_ptr<class_type>;
-	using map = std::map<std::string, ptr>;
-	virtual bool matches(const std::string& value, const map& type_map) const = 0;
-	static std::shared_ptr<class_type> from_config(const config& cfg);
-};
-
-/**
- * Stores information about a schema type.
- * This type represents a simple pattern match.
- */
-class class_type_simple : public class_type {
-	boost::regex pattern_;
-public:
-	class_type_simple(const std::string& name, const std::string& pattern) : class_type(name), pattern_(pattern) {}
-	bool matches(const std::string& value, const map& type_map) const override;
-};
-
-/**
- * Stores information about a schema type.
- * This type represents a name alias for another type.
- */
-class class_type_alias : public class_type {
-	mutable std::shared_ptr<class_type> cached_;
-	std::string link_;
-public:
-	class_type_alias(const std::string& name, const std::string& link) : class_type(name), link_(link) {}
-	bool matches(const std::string& value, const map& type_map) const override;
-};
-
-/**
- * Stores information about a schema type.
- * This is an abstract base class for composite types.
- */
-class class_type_composite : public class_type {
-protected:
-	std::vector<std::shared_ptr<class_type>> subtypes_;
-public:
-	explicit class_type_composite(const std::string& name) : class_type(name) {}
-	void add_type(std::shared_ptr<class_type> type)
-	{
-		subtypes_.push_back(type);
-	}
-};
-
-/**
- * Stores information about a schema type.
- * Represents a union type, which matches if any of its subtypes match.
- */
-class class_type_union : public class_type_composite {
-public:
-	explicit class_type_union(const std::string& name) : class_type_composite(name) {}
-	bool matches(const std::string& value, const map& type_map) const override;
-};
-
-/**
- * Stores information about a schema type.
- * Represents an intersection type, which matches if all of its subtypes match.
- */
-class class_type_intersection : public class_type_composite {
-public:
-	explicit class_type_intersection(const std::string& name) : class_type_composite(name) {}
-	bool matches(const std::string& value, const map& type_map) const override;
-};
-
-/**
- * Stores information about a schema type.
- * Represents a list type, where each list element is itself a union.
- */
-class class_type_list : public class_type_union {
-	boost::regex split_;
-	int min_ = 0, max_ = -1;
-public:
-	class_type_list(const std::string& name, const std::string& pattern, int min, int max)
-		: class_type_union(name)
-		, split_(pattern)
-		, min_(min)
-		, max_(max)
-	{}
-	bool matches(const std::string& value, const map& type_map) const override;
-};
-
-/**
- * class_key is used to save the information about one key.
- * Key has next info: name, type, default value or key is mandatory.
- */
-class class_key
-{
-public:
-	class_key()
-		: name_("")
-		, type_("")
-		, default_("\"\"")
-		, mandatory_(false)
-		, fuzzy_(false)
-	{
-	}
-
-	class_key(const std::string& name, const std::string& type, const std::string& def = "\"\"")
-		: name_(name)
-		, type_(type)
-		, default_(def)
-		, mandatory_(def.empty())
-		, fuzzy_(name.find_first_of("*?") != std::string::npos)
-	{
-	}
-
-	class_key(const config&);
-
-	const std::string& get_name() const
-	{
-		return name_;
-	}
-
-	const std::string& get_type() const
-	{
-		return type_;
-	}
-
-	const std::string& get_default() const
-	{
-		return default_;
-	}
-
-	bool is_mandatory() const
-	{
-		return mandatory_;
-	}
-
-	bool is_fuzzy() const {
-		return fuzzy_;
-	}
-
-	void set_name(const std::string& name)
-	{
-		name_ = name;
-	}
-
-	void set_type(const std::string& type)
-	{
-		type_ = type;
-	}
-
-	void set_default(const std::string& def)
-	{
-		default_ = def;
-		if(def.empty()) {
-			mandatory_ = true;
-		}
-	}
-
-	void set_mandatory(bool mandatory)
-	{
-		mandatory_ = mandatory;
-	}
-
-	void set_fuzzy(bool f)
-	{
-		fuzzy_ = f;
-	}
-
-	/** is used to print key info
-	 * the format is next
-	 *  [key]
-	 *      name="name"
-	 *      type="type"
-	 *      default="default"
-	 *      mandatory="true/false"
-	 *  [/key]
-	 */
-	void print(std::ostream& os, int level) const;
-
-	/** Compares keys by name. Used in std::sort, i.e. */
-	bool operator<(const class_key& k) const
-	{
-		return (get_name() < k.get_name());
-	}
-
-private:
-	/** Name of key. */
-	std::string name_;
-
-	/** Type of key. */
-	std::string type_;
-
-	/** Default value. */
-	std::string default_;
-
-	/** Shows, if key is a mandatory key. */
-	bool mandatory_;
-
-	/** Whether the key is a fuzzy match. */
-	bool fuzzy_;
-};
 
 class class_condition;
 
@@ -413,21 +206,8 @@ public:
 		max_ = o;
 	}
 
-	void set_min(const std::string& s)
-	{
-		std::istringstream i(s);
-		if(!(i >> min_)) {
-			min_ = 0;
-		}
-	}
-
-	void set_max(const std::string& s)
-	{
-		std::istringstream i(s);
-		if(!(i >> max_)) {
-			max_ = 0;
-		}
-	}
+	void set_min(const std::string& s);
+	void set_max(const std::string& s);
 
 	void set_super(const std::string& s)
 	{
